refactor: Change Cover to ChartClip
