+++ packages/core/src/charts/area-stacked.ts
@@ -6,7 +6,7 @@ import { Tools } from "../tools";
 
 // Components
 import {
-	Cover,
+	ChartClip,
 	Grid,
 	StackedArea,
 	TwoDimensionalAxes,
@@ -36,7 +36,7 @@ export class StackedAreaChart extends AxisChart {
 		// Specify what to render inside the graph-frame
 		const graphFrameComponents: any[] = [
 			new TwoDimensionalAxes(this.model, this.services),
-			new Cover(this.model, this.services),
+			new ChartClip(this.model, this.services),
 			new Grid(this.model, this.services),
 			new Ruler(this.model, this.services),
 			new StackedArea(this.model, this.services),
