+++ packages/core/src/charts/area.ts
@@ -6,7 +6,7 @@ import { Tools } from "../tools";
 
 // Components
 import {
-	Cover,
+	ChartClip,
 	Grid,
 	Area,
 	Line,
@@ -40,7 +40,7 @@ export class AreaChart extends AxisChart {
 		// Specify what to render inside the graph-frame
 		const graphFrameComponents: any[] = [
 			new TwoDimensionalAxes(this.model, this.services),
-			new Cover(this.model, this.services),
+			new ChartClip(this.model, this.services),
 			new Grid(this.model, this.services),
 			new Ruler(this.model, this.services),
 			new Line(this.model, this.services),
