+++ packages/core/src/charts/bar-stacked.ts
@@ -7,7 +7,7 @@ import { Skeletons } from "../interfaces/enums";
 
 // Components
 import {
-	Cover,
+	ChartClip,
 	Grid,
 	StackedBar,
 	StackedBarRuler,
@@ -43,7 +43,7 @@ export class StackedBarChart extends AxisChart {
 		// Specify what to render inside the graph-frame
 		const graphFrameComponents: any[] = [
 			new TwoDimensionalAxes(this.model, this.services),
-			new Cover(this.model, this.services),
+			new ChartClip(this.model, this.services),
 			new Grid(this.model, this.services),
 			new StackedBarRuler(this.model, this.services),
 			new StackedBar(this.model, this.services),
