+++ packages/core/src/charts/line.ts
@@ -7,7 +7,7 @@ import { Skeletons } from "../interfaces/enums";
 
 // Components
 import {
-	Cover,
+	ChartClip,
 	Grid,
 	Line,
 	Ruler,
@@ -41,7 +41,7 @@ export class LineChart extends AxisChart {
 		// Specify what to render inside the graph-frame
 		const graphFrameComponents: any[] = [
 			new TwoDimensionalAxes(this.model, this.services),
-			new Cover(this.model, this.services),
+			new ChartClip(this.model, this.services),
 			new Grid(this.model, this.services),
 			new Ruler(this.model, this.services),
 			new Line(this.model, this.services),
